Use explicit `static_cast` for type conversion
I assume the casting to `int` was to force rounding. Either that, or the
original cast was accidental and kind of useless. At any rate, using a
`static_cast` for both directions should preserve the previous
behavior.

+++ NAS2D/Renderer/RendererOpenGL.cpp
@@ -819,7 +819,7 @@ namespace
 
 		float t = 0.0f;
 		float R = 0.0f;
-		float f = lineWidth - static_cast<int>(lineWidth);
+		float f = lineWidth - static_cast<float>(static_cast<int>(lineWidth));
 
 		if (lineWidth >= 0.0f && lineWidth < 1.0f)
 		{
